requirements.txt
keras==2.9.0
numpy==1.23.5
opencv_python_headless==4.10.0.84
Pillow==11.1.0
streamlit==1.20.0
